pyproject.toml
[tool.poetry]
name = "weirdtext-api"
version = "0.1.0"
description = ""
authors = ["szymos <skoczylas9szymon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
pytest = "^6.2.5"
django-rest = "^0.8.7"
djangorestframework = "^3.12.4"
django-pytest = "^0.2.0"
weirdtext-lib = {git = "https://github.com/szymos/WeirdText-lib.git"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
